+++ ducktape/cluster/vagrant.py
